Cart

+++ controllers/orderControllers.js
@@ -6,7 +6,7 @@ exports.getallOrder = async (req, res)=>{
         users.u_userName,
         usersdetail.de_firstName,usersdetail.de_lastName,usersdetail.de_tel 
         FROM orders INNER JOIN users ON users.u_ID = orders.u_ID
-        INNER JOIN usersdetail ON usersdetail.u_ID = orders.u_ID
+        INNER JOIN usersdetail ON usersdetail.u_ID = orders.u_ID ORDER BY orders.o_ID ASC
         `
         const [orderResult] = await conn.query(orderSQL)
         if(orderResult.length === 0 ){
@@ -61,7 +61,7 @@ exports.getOrderID= (req, res)=>{
 }
 
 exports.addOrder = async (req,res) => {
-    const {cart,u_ID,o_endDate} = req.body
+    const {cart,o_endDate,u_ID} = req.body
     const io = req.app.get('io')
     try {
         if(!u_ID ||  !o_endDate) {
@@ -75,7 +75,7 @@ exports.addOrder = async (req,res) => {
         const o_ID = orderResult.insertId
         for ( const item of cart) {
             const orderItemSQL = `INSERT INTO ordersitems (o_ID,p_ID,i_Amount ) VALUE (?,?,?) `
-            const [orderItemResult] = await conn.query(orderItemSQL,[o_ID,item.p_ID,item.p_Amount,item.p_Price])
+            const [orderItemResult] = await conn.query(orderItemSQL,[o_ID,item.p_ID,item.i_Amount,item.p_Price])
             if (orderItemResult.affectedRows === 0) {
                 return res.status(400).send({message : `Can't Insert Order items` , status : 0})
             }
@@ -164,6 +164,9 @@ exports.deleteOrders = async (req,res)=>{
       }
       const deleteOrderItemSQL = `DELETE FROM ordersitems WHERE o_ID = ?`
       const [resultdeleteOrderItem] = await conn.query(deleteOrderItemSQL,[o_ID])
+      if (resultdeleteOrder.affectedRows === 0){
+        return res.status(404).send({message: `Delete Order Items Unsuccessfully`, status : 0})
+      }
       const deleteOrderSQL = `DELETE FROM orders WHERE o_ID = ?`
       const [resultdeleteOrder] = await conn.query(deleteOrderSQL,[o_ID])
       if (resultdeleteOrder.affectedRows === 0) {
@@ -171,6 +174,29 @@ exports.deleteOrders = async (req,res)=>{
       } 
       io.emit('refreshOrders')
       return res.status(200).send({message : `Delete Order ID : ${o_ID} Successfully`, status : 1})
+    } catch (error) {
+        console.log(error)
+        return res.status(500).send({message : `Somethings Went Wrong` , status : 0})
+    }
+}
+
+
+exports.sumPrice = async(req,res)=>{
+    const {o_ID} = req.params
+    let sum = 0
+    try {
+        if(!o_ID) {
+            return res.status(400).send({message: `Order ID is Missing`, status: 0})
+        }
+        const selectItemAmountSQL= `SELECT product.p_Price,ordersitems.p_ID,ordersitems.i_Amount FROM ordersitems 
+                                    INNER JOIN product ON product.p_ID = ordersitems.p_ID WHERE o_ID = ?`
+        const [selectItemAmount] = await conn.query(selectItemAmountSQL,[o_ID])
+        for (const item of selectItemAmount){
+            sum += item.p_Price * item.i_Amount
+            // console.log(`p_ID : ${item.p_ID}  price : ${item.p_Price}  Amount : ${item.i_Amount}`)
+        }
+        const data = sum
+        return res.status(200).send({message : `Sum success`, data ,status : 1})
     } catch (error) {
         console.log(error)
         return res.status(500).send({message : `Somethings Went Wrong` , status : 0})
